Modularize prefixes and selection
