+++ helpers/prefix_good.py
@@ -1,0 +1,2 @@
+# Präfix für als gut bewertete Marker
+PREFIX_GOOD = "GOOD_"
