+++ helpers/prefix_new.py
@@ -1,0 +1,2 @@
+# Präfix für neu erkannte Tracks
+PREFIX_NEW = "NEW_"
