+++ helpers/prefix_track.py
@@ -1,0 +1,2 @@
+# Präfix für aktive Tracking-Marker
+PREFIX_TRACK = "TRACK_"
